refactor: rename schema.ts to patients.ts
