+++ src/db/schema/auth.ts
@@ -1,7 +1,4 @@
-import {
-    pgSchema,
-    uuid,
-} from "drizzle-orm/pg-core";
+import { pgSchema, uuid } from "drizzle-orm/pg-core";
 
 // Stub for Supabase-managed auth schema — not migrated, used only for FK references
 const authSchema = pgSchema("auth");
